savepoint
it's all good for that moment, but still need to:
1) Implement Deleting rows
2) Refactor notifications logic
